Initial commit of the new method "prayer"
The new method is a "clean-room" implementation of the prayer times
and qibla direction calculations. This new implementation uses
simpler but approximate methods. This should make the code much
simpler and easier to maintain. The impact on accuracy is still
to be determined. Please NOTE that this is still WIP and far from
being complete