Change threshold to 4 minutes in run_tests
I changed the altitude in the configuration of Mecca and Eindhoven.
The altitude of Mecca is set now to the altitude in Al-Masjid Al-Haram
(~ 300 m) and for Eindhoven it is around 21 m. We need to see the reason
for the increased difference between the Official Umm Alqura values and
the computed values. I suspect that for Um Alqura, they use and altitude
equal to the sea level.